Split Experiment into modules
