+++ marconi-core/src/Marconi/Core/Experiment/Transformer/WithCache.hs
@@ -1,0 +1,203 @@
+{-# LANGUAGE DerivingVia          #-}
+{-# LANGUAGE StrictData           #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-redundant-constraints #-}
+
+{- |
+    A transformer that cache result of some queries
+
+    See "Marconi.Core.Experiment" for documentation.
+ -}
+module Marconi.Core.Experiment.Transformer.WithCache
+    ( WithCache
+        , withCache
+        , addCacheFor
+        , cachedIndexer
+    ) where
+
+import Control.Lens (At (at), Getter, Indexable (indexed), IndexedTraversal', Lens', itraverseOf, lens, makeLenses, to)
+import Control.Lens.Operators ((%~), (&), (^.))
+import Control.Monad.Except (ExceptT, MonadError (throwError), runExceptT)
+import Data.Foldable (foldl')
+import Data.Map (Map)
+import Data.Map qualified as Map
+
+import Marconi.Core.Experiment.Class (Closeable, HasGenesis (genesis), IsIndex (index, indexAll), IsSync,
+                                      Queryable (query), Resetable (reset), Rollbackable (rollback))
+import Marconi.Core.Experiment.Transformer.IndexWrapper (IndexWrapper (IndexWrapper), indexAllVia, indexVia,
+                                                         lastSyncPointVia, queryLatestVia, queryVia, resetVia,
+                                                         rollbackVia, wrappedIndexer, wrapperConfig)
+import Marconi.Core.Experiment.Type (IndexerError (OtherIndexError), Point, QueryError (AheadOfLastSync), Result,
+                                     TimedEvent)
+
+data CacheConfig query event
+    = CacheConfig
+      { _configCache     :: Map query (Result query)
+      , _configOnForward :: TimedEvent event -> Result query -> Result query
+      }
+
+configCache :: Lens' (CacheConfig query event) (Map query (Result query))
+configCache = lens _configCache (\cfg c -> cfg {_configCache = c})
+
+configCacheEntries :: IndexedTraversal' query (CacheConfig query event) (Result query)
+configCacheEntries f cfg
+    = (\c -> cfg {_configCache = c})
+    <$> Map.traverseWithKey (indexed f) (_configCache cfg)
+
+configOnForward
+    :: Getter
+        (CacheConfig query event)
+        (TimedEvent event -> Result query -> Result query)
+configOnForward = to _configOnForward
+
+-- | Setup a cache for some requests.
+--
+-- The cache is active only for the latest `Point`.
+-- As a consequence, using `WithCache` is more effective on top of the on-disk
+-- part of a `MixedIndexer`, or any other part of an indexer that has a relatively
+-- stable sync point.
+newtype WithCache query indexer event
+    = WithCache { _cacheWrapper :: IndexWrapper (CacheConfig query) indexer event }
+
+makeLenses 'WithCache
+
+deriving via (IndexWrapper (CacheConfig query) indexer)
+    instance IsSync m event indexer => IsSync m event (WithCache query indexer)
+
+deriving via (IndexWrapper (CacheConfig query) indexer)
+    instance Closeable m indexer => Closeable m (WithCache query indexer)
+
+-- | A smart constructor for 'WithCache'.
+-- The cache starts empty, you can populate it with 'addCacheFor'
+withCache
+    :: Ord query
+    => (TimedEvent event -> Result query -> Result query)
+    -> indexer event
+    -> WithCache query indexer event
+withCache _configOnForward
+    = WithCache . IndexWrapper
+    ( CacheConfig
+        { _configCache = mempty
+        , _configOnForward
+        }
+    )
+
+-- | A (indexed-)traversal to all the entries of the cache
+cacheEntries :: IndexedTraversal' query (WithCache query indexer event) (Result query)
+cacheEntries = cacheWrapper . wrapperConfig . configCacheEntries
+
+-- | Access to the cache
+cache :: Lens' (WithCache query indexer event) (Map query (Result query))
+cache = cacheWrapper . wrapperConfig . configCache
+
+-- | How do we add event to existing cache
+onForward
+    :: Getter
+        (WithCache query indexer event)
+        (TimedEvent event -> Result query -> Result query)
+onForward = cacheWrapper . wrapperConfig . configOnForward
+
+-- | Access to the indexer that is cached.
+cachedIndexer :: Lens' (WithCache query indexer event) (indexer event)
+cachedIndexer = cacheWrapper . wrappedIndexer
+
+-- | Add a cache for a specific query.
+--
+-- When added, the cache query the underlying indexer to populate the cache for this query.
+--
+-- If you want to add several indexers at the same time, use traverse.
+addCacheFor
+    :: Queryable (ExceptT (QueryError query) m) event query indexer
+    => IsSync (ExceptT (QueryError query) m) event indexer
+    => Monad m
+    => MonadError IndexerError m
+    => Ord query
+    => Ord (Point event)
+    => query
+    -> WithCache query indexer event
+    -> m (WithCache query indexer event)
+addCacheFor q indexer
+    = do
+        initialResult <- runExceptT $ queryLatestVia cachedIndexer q indexer
+        case initialResult of
+            Left _err    -> throwError $ OtherIndexError "Can't create cache"
+            Right result -> pure $ indexer & cache %~ Map.insert q result
+
+-- | This instances update all the cached queries with the incoming event
+-- and then pass this event to the underlying indexer.
+instance
+    (Applicative m, IsIndex m event index)
+    => IsIndex m event (WithCache query index) where
+
+    index timedEvent indexer = do
+        indexer' <- indexVia cachedIndexer timedEvent indexer
+        pure $ indexer' & cacheEntries %~ (indexer' ^. onForward) timedEvent
+
+    indexAll evts indexer = do
+        indexer' <- indexAllVia cachedIndexer evts indexer
+        pure $ indexer' & cacheEntries %~ flip (foldl' (flip $ indexer' ^. onForward)) evts
+
+rollbackCache
+    :: Applicative f
+    => Ord (Point event)
+    => Queryable f event query indexer
+    => Point event
+    -> WithCache query indexer event
+    -> f (WithCache query indexer event)
+rollbackCache p indexer
+    = itraverseOf
+        cacheEntries
+        (\q -> const $ queryVia cachedIndexer p q indexer)
+        indexer
+
+-- | Rollback the underlying indexer, clear the cache,
+-- repopulate it with queries to the underlying indexer.
+instance
+    ( Monad m
+    , Rollbackable m event index
+    , HasGenesis (Point event)
+    , Queryable m event query index
+    ) => Rollbackable m event (WithCache query index) where
+
+    rollback p indexer = do
+        res <- rollbackVia cachedIndexer p indexer
+        rollbackCache p res
+
+-- | Rollback the underlying indexer, clear the cache,
+-- repopulate it with queries to the underlying indexer.
+instance
+    ( Monad m
+    , Resetable m event index
+    , HasGenesis (Point event)
+    , Ord (Point event)
+    , Queryable m event query index
+    ) => Resetable m event (WithCache query index) where
+
+    reset indexer = do
+        res <- resetVia cachedIndexer indexer
+        rollbackCache genesis res
+
+instance
+    (Ord query
+    , Ord (Point event)
+    , IsSync m event index
+    , MonadError (QueryError query) m
+    , Monad m
+    , Queryable m event query index
+    ) => Queryable m event query (WithCache query index) where
+
+
+    -- | On a query, if it miss the cache, query the indexer.
+    -- If the cache is fresher than the request point, query the underlying indexer,
+    -- if it's the cached point, send the result.
+    -- If the cache is behind the requested point, send an 'AheadOfLastSync' error
+    -- with the cached content.
+    query p q indexer = do
+        syncPoint' <- lastSyncPointVia cachedIndexer indexer
+        let cached = indexer ^. cache . at q
+        let queryWithoutCache = queryVia cachedIndexer p q indexer
+        case compare p syncPoint' of
+            LT -> queryWithoutCache
+            EQ -> maybe queryWithoutCache pure cached
+            GT -> maybe queryWithoutCache (throwError . AheadOfLastSync . Just) cached
+
